strategify-backend/requirements.txt
pyrebase
websocket-client
numpy